Update usachvirtualVectorial1.c
- Se arreglo en linea 38 por incompatibilidad de 'int' con 'size_t'

+++ usachvirtualVectorial1.c
@@ -1,4 +1,3 @@
-#include <xmmintrin.h>
 #include <emmintrin.h>
 #include <pmmintrin.h>
 #include <math.h>
@@ -36,7 +35,7 @@ int main(int argc, char *argv[])
 					 N = N-1;
 					int j=0;
 					 //float a[N] __attribute__((aligned(16)));
-					    size_t t[N];
+					    int t[N];
 					    FILE *fichero1;
 					    fichero1 = fopen(nombre,"r");
 					    if (fichero1 == NULL){
